+++ src/components/Home/HomeHistory.tsx
@@ -14,7 +14,7 @@ import { twMerge } from 'tailwind-merge'
 import Card, { CardPosition, getCardPosition } from '../Global/Card'
 import EmptyState from '../Global/EmptyStates/EmptyState'
 import { KycStatusItem } from '../Kyc/KycStatusItem'
-import { isKycStatusItem, KycHistoryEntry } from '@/hooks/useKycFlow'
+import { isKycStatusItem, KycHistoryEntry } from '@/hooks/useBridgeKycFlow'
 import { BridgeKycStatus } from '@/utils'
 import { useWallet } from '@/hooks/wallet/useWallet'
 import { useUserInteractions } from '@/hooks/useUserInteractions'
